+++ app/src/main/java/space/taran/arknavigator/di/modules/RepoModule.kt
@@ -8,7 +8,7 @@ import space.taran.arknavigator.mvp.model.dao.Database
 import space.taran.arknavigator.mvp.model.repo.FoldersRepo
 import space.taran.arknavigator.mvp.model.repo.index.ResourcesIndexRepo
 import space.taran.arknavigator.mvp.model.repo.tags.TagsStorageRepo
-import space.taran.arknavigator.utils.MAIN
+import space.taran.arknavigator.utils.LogTags.MAIN
 import javax.inject.Singleton
 
 @Module
