+++ app/src/main/java/space/taran/arknavigator/mvp/model/repo/preview/PreviewGenerators.kt
@@ -11,7 +11,7 @@ import space.taran.arknavigator.mvp.model.repo.extra.ImageMetaExtra
 import space.taran.arknavigator.mvp.model.repo.preview.generator.PdfPreviewGenerator
 import space.taran.arknavigator.ui.App
 import space.taran.arknavigator.utils.ImageUtils.glideExceptionListener
-import space.taran.arknavigator.utils.PREVIEWS
+import space.taran.arknavigator.utils.LogTags.PREVIEWS
 import space.taran.arknavigator.utils.extension
 import java.nio.file.Files
 import java.nio.file.Path
