+++ app/src/main/java/space/taran/arknavigator/utils/FamilyApps.kt
@@ -1,0 +1,5 @@
+package space.taran.arknavigator.utils
+
+object FamilyApps {
+    const val RETOUCH: String = "space.taran.arkretouch"
+}
