+++ app/src/main/java/space/taran/arknavigator/utils/LogTags.kt
@@ -1,40 +1,42 @@
 package space.taran.arknavigator.utils
 
-val Any.TAG: String
-    get() = this::class.java.simpleName
+// val Any.TAG: String
+//     get() = this::class.java.simpleName
 
-const val MAIN: String = "main"
+object LogTags {
+    const val MAIN: String = "main"
 
-const val MODULES: String = "modules"
+    const val MODULES: String = "modules"
 
-const val PERMISSIONS: String = "permissions"
+    const val PERMISSIONS: String = "permissions"
 
-const val RESOURCES_SCREEN: String = "resources-screen"
+    const val RESOURCES_SCREEN: String = "resources-screen"
 
-const val RESOURCES_INDEX: String = "resources-index"
+    const val RESOURCES_INDEX: String = "resources-index"
 
-const val GALLERY_SCREEN: String = "gallery-screen"
+    const val GALLERY_SCREEN: String = "gallery-screen"
 
-const val FOLDERS_SCREEN: String = "folders-screen"
+    const val FOLDERS_SCREEN: String = "folders-screen"
 
-const val SETTINGS_SCREEN: String = "settings-screen"
+    const val SETTINGS_SCREEN: String = "settings-screen"
 
-const val FOLDER_PICKER: String = "folder-picker"
+    const val FOLDER_PICKER: String = "folder-picker"
 
-const val FOLDERS_TREE: String = "folders-tree"
+    const val FOLDERS_TREE: String = "folders-tree"
 
-const val TAGS_SELECTOR: String = "tags-selector"
+    const val TAGS_SELECTOR: String = "tags-selector"
 
-const val TAGS_STORAGE: String = "tags-storage"
+    const val TAGS_STORAGE: String = "tags-storage"
 
-const val ITEMS_CONTAINER: String = "item-container"
+    const val ITEMS_CONTAINER: String = "item-container"
 
-const val DATABASE: String = "database"
+    const val DATABASE: String = "database"
 
-const val PREVIEWS: String = "previews"
+    const val PREVIEWS: String = "previews"
 
-const val IMAGES: String = "images"
+    const val IMAGES: String = "images"
 
-const val GLIDE: String = "glide"
+    const val GLIDE: String = "glide"
 
-const val FILES: String = "files"
+    const val FILES: String = "files"
+}
